requirements.txt
fastapi
uvicorn
qrcode
pillow
python-multipart